refactor(processor): extract pdfjs warning capture
Move pdf.js warning interception into a small helper so the document processor keeps less global console plumbing inline.
